requirements.txt
flask==2.3.3
openai-whisper==20231117
python-dotenv==1.0.0
werkzeug==2.3.7
torch
torchaudio
ffmpeg-python 